refactor statesman
statesman might actually work now! the roots are now collected in a jquery object so iteration is easier, and the iteration itself was fixed because it was mad stupid

next up for statesman is an exception to the hidden class for main pages. so enjoy this before i ruin it again